Check in a glibc 2.17 + gcc 4.8.3 toolchain
This CL checks in glibc 2.17. This was built using the newly-included
build-with-previous-gcc.sh script. build-*-multilib-toolchain.sh is more
or less the same as it was; it learned to copy its wrapper into the new
toolchain root, and now knows how to apply things in sysroot-patches.
For a full diff between old and new, look here (sorry external people;
please use `diff` manually):
https://paste.googleplex.com/4796800347144192

In the 2.15 upgrade CL (I355e28b922101700719c0cc27f588782f96979d4), the
commit message notes that it also:
- added a symbolic link to stdatomic.h
- copied libxcb.so.1{,.1.0} from precise's /usr/lib/* to the prebuilts
  glibc dir.

Despite no builds breaking without it, I added the first bullet as a
sysroot-patch.

For the second bullet, it seems like a really bad idea to try to copy
libxcb.so from a modernish Debian into an essentially 5 year old
sysroot. cs/ for `libxcb package:^android$ -file:^prebuilts/
-file:quipper/testdata -file:^external/` gives me a single result.
Namely, a directory in vendor/ apparently provides its own libxcb.so.1.
(And a few places in external/ reference it in documentation? Hard to
tell if we ever actually need it, though)

So, unless we can find something it breaks, I'm dropping the libxcb bits
from this (note that the sysroot *does* still contain a libxcb on its
own, like the previous one does.)

Manual actions:
  - Copied over CAPS root files from the glibc-2.15 dir not already
    present in this directory (MODULE_LICENSE_GPL, OWNERS, NOTICE, etc.)

Bug: 111358231
Test: A few:
  - checkbuilds of taimen-eng, hikey960-eng, marlin-eng,
    sdk_gphone_x86-userdebug
  - `make -j80 dist TARGET_PRODUCT=cf_x86_phone ndk_translation_all`
  - `emulator -no-display` on sdk_gphone_x86-userdebug booted and was
    `adb`able
  - built the emulator branch
    (I8936e4a3bddfd500dadb8034f0b272eb1c463978); tests pass. Using the
    emulator on a newly-built sdk_gphone_x86-userdebug image boots and
    is `adb`able
  - `make -j80 -k PRODUCT-sdk_x86-sdk dist` builds

Change-Id: Ieeed444801160e3d288bbad7d871067e81411118